refactor[api-server]: lottery.dto 패키지 위치 변경
