+++ src/main/java/com/watermelon/server/event/lottery/domain/Expectation.java
@@ -1,7 +1,7 @@
 package com.watermelon.server.event.lottery.domain;
 
 import com.watermelon.server.BaseEntity;
-import com.watermelon.server.lottery.dto.request.RequestExpectationDto;
+import com.watermelon.server.event.lottery.dto.request.RequestExpectationDto;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
