+++ src/main/java/com/watermelon/server/event/lottery/dto/response/ResponseLotteryWinnerDto.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.lottery.dto.response;
+package com.watermelon.server.event.lottery.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
